requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
jobspy==0.31.0
anthropic==0.34.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
PyPDF2==3.0.1
python-dotenv==1.0.0
Werkzeug==2.3.7
python-docx==1.1.0
lxml==4.9.3